Correctif pour apk téléphone Erreur lors de la compilation: The following ArgumentError was thrown resolving an image codec: Invalid argument(s): No host specified in URI file:///Chargement%20...
The following assertion was thrown during layout:
A RenderFlex overflowed by 132 pixels on the right.
The relevant error-causing widget was:
  Row Row:file:///C:/Users/julie.pouny/Documents/Projets/Revision%20Flutter/app_meteo/lib/widgets/home.dart:137:25
The overflowing RenderFlex has an orientation of Axis.horizontal.
The edge of the RenderFlex that is overflowing has been marked in the rendering with a yellow and black striped pattern. This is usually caused by the contents being too big for the RenderFlex

The following assertion was thrown during layout:
A RenderFlex overflowed by 5050 pixels on the bottom.
The relevant error-causing widget was:
  Column Column:file:///C:/Users/julie.pouny/Documents/Projets/Revision%20Flutter/app_meteo/lib/widgets/home.dart:119:18
The overflowing RenderFlex has an orientation of Axis.vertical.
The edge of the RenderFlex that is overflowing has been marked in the rendering with a yellow and black striped pattern. This is usually caused by the contents being too big for the RenderFlex.